+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Breaking API Change: provide labels when intersection is found: `TBLError::Intersection(Option<L>, Option<L>)`.
+
 ## [1.0.0-alpha](https://crates.io/crates/tbl/1.0.0-alpha) Jun 13, 2020
 
 * add multiline, `Renderer::render` now returns `Vec<String>`.